fix: patch refresh token logic
