+++ apps/web/src/providers/ModalProvider.tsx
@@ -1,4 +1,4 @@
-import { useAtom, useAtomValue, useSetAtom } from 'jotai';
+import { useAtom, useAtomValue } from 'jotai';
 import dynamic from 'next/dynamic';
 import { useRouter } from 'next/router';
 import type React from 'react';
@@ -6,16 +6,15 @@ import { useCallback } from 'react';
 
 import {
   currentWorkspaceIdAtom,
-  jotaiWorkspacesAtom,
   openCreateWorkspaceModalAtom,
   openWorkspacesModalAtom,
 } from '../atoms';
 import { useCurrentUser } from '../hooks/current/use-current-user';
 import { useCurrentWorkspace } from '../hooks/current/use-current-workspace';
+import { useOnGoogleLogout } from '../hooks/use-on-google-logout';
 import { useRouterHelper } from '../hooks/use-router-helper';
 import { useWorkspaces, useWorkspacesHelper } from '../hooks/use-workspaces';
-import { WorkspacePlugins } from '../plugins';
-import { RemWorkspaceFlavour, WorkspaceSubPath } from '../shared';
+import { WorkspaceSubPath } from '../shared';
 import { apis } from '../shared/apis';
 
 const WorkspaceListModal = dynamic(
@@ -43,7 +42,6 @@ export function Modals() {
   const currentWorkspaceId = useAtomValue(currentWorkspaceIdAtom);
   const [, setCurrentWorkspace] = useCurrentWorkspace();
   const { createLocalWorkspace } = useWorkspacesHelper();
-  const set = useSetAtom(jotaiWorkspacesAtom);
 
   return (
     <>
@@ -76,16 +74,7 @@ export function Modals() {
             router.reload();
           });
         }, [router])}
-        onClickLogout={useCallback(() => {
-          apis.auth.clear();
-          set(workspaces =>
-            workspaces.filter(
-              workspace => workspace.flavour !== RemWorkspaceFlavour.AFFINE
-            )
-          );
-          WorkspacePlugins[RemWorkspaceFlavour.AFFINE].cleanup?.();
-          router.reload();
-        }, [router, set])}
+        onClickLogout={useOnGoogleLogout()}
         onCreateWorkspace={useCallback(() => {
           setOpenCreateWorkspaceModal(true);
         }, [setOpenCreateWorkspaceModal])}
